second commit

+++ internal/service/task_service.go
@@ -1,44 +1,51 @@
 package service
 
 import (
-	"manager/internal/model"
-	"manager/internal/repository"
+  "errors"
+  "strings"
+  "manager/internal/model"
+  "manager/internal/repository"
 )
 
-type ITaskService interface {
-	Create(task *model.Task) error
-	GetAll() ([]model.Task, error)
-	GetByID(id uint) (*model.Task, error)
-	Update(task *model.Task) error
-	Delete(id uint) error
+type ItaskService interface {
+  Create(task model.Task) error
+  GetAll() ([]model.Task, error)
+  GetByID(id int) (*model.Task, error)
+  Delete(id int) error
 }
 
 type service struct {
-	repo repository.ITaskRepo
+  repo repository.ITaskRepo
 }
 
-func New(repo repository.ITaskRepo) ITaskService {
-	return &service{
-		repo: repo,
-	}
+func New(repo repository.ITaskRepo) ItaskService {
+  return &service{
+    repo: repo,
+  }
 }
 
-func (s *service) Create(task *model.Task) error {
-	return s.repo.Create(task)
+func (s *service) Create(task model.Task) error {
+  if strings.TrimSpace(task.Title) == "" {
+    return errors.New("empty title")
+  }
+
+  return s.repo.Create(&task)
 }
 
 func (s *service) GetAll() ([]model.Task, error) {
-	return s.repo.GetAll()
+  return s.repo.GetAll()
 }
 
-func (s *service) GetByID(id uint) (*model.Task, error) {
-	return s.repo.GetByID(id)
+func (s *service) GetByID(id int) (*model.Task, error) {
+  if id == 0 {
+    return nil, errors.New("invalid id")
+  }
+  return s.repo.GetByID(id)
 }
 
-func (s *service) Update(task *model.Task) error {
-	return s.repo.Update(task)
-}
-
-func (s *service) Delete(id uint) error {
-	return s.repo.Delete(id)
+func (s *service) Delete(id int) error {
+  if id == 0 {
+    return errors.New("invalid id")
+  }
+  return s.repo.Delete(id)
 }
